requirements/base.txt
Django==1.5.1
bpython==0.12
django-model-utils==1.4.0
logutils==0.3.3
South==0.8.1
requests
redis==2.9.1
celery==3.1.11
MySQL-python==1.2.5
django-extensions==1.2.2
ipython
django-allauth==0.17.0

# Forms
django-braces==1.4.0
django-crispy-forms==1.4.0
django-floppyforms==1.1.1

django-avatar==2.0
python-instagram==1.1.2
